third_party/blink/web_tests/platform/win11-arm64/http/tests/security/isolatedWorld/isolated-world-fetch-csp-expected.txt

CONSOLE MESSAGE: Testing main world. Request should be blocked by main world CSP.
CONSOLE ERROR: Refused to connect to 'http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php' because it violates the following Content Security Policy directive: "connect-src 'none'".

CONSOLE ERROR: Refused to connect to 'http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php' because it violates the document's Content Security Policy.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
CONSOLE MESSAGE: Testing isolated world with no csp. Request should be blocked by main world CSP.
CONSOLE ERROR: Refused to connect to 'http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php' because it violates the following Content Security Policy directive: "connect-src 'none'".

CONSOLE ERROR: Refused to connect to 'http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php' because it violates the document's Content Security Policy.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
CONSOLE MESSAGE: Testing isolated world with permissive csp.
CONSOLE MESSAGE: PASS: Request succeeded as expected.
CONSOLE MESSAGE: Testing fetch redirect in isolated world with permissive csp.
CONSOLE MESSAGE: FAIL: Request succeeded with incorrect response <br />
<b>Warning</b>:  Undefined array key "refresh" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\resources\redirect.php</b> on line <b>6</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\resources\redirect.php:6) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\resources\redirect.php</b> on line <b>14</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\resources\redirect.php:6) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\resources\redirect.php</b> on line <b>16</b><br />
<br />
<b>Warning</b>:  Undefined array key "code" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\resources\redirect.php</b> on line <b>20</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\resources\redirect.php:6) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\resources\redirect.php</b> on line <b>23</b><br />

CONSOLE MESSAGE: Testing isolated world with strict csp.
CONSOLE ERROR: Refused to connect to 'http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php' because it violates the following Content Security Policy directive: "connect-src 'self'".

CONSOLE ERROR: Refused to connect to 'http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php' because it violates the document's Content Security Policy.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
CONSOLE MESSAGE: Testing fetch redirect in isolated world with strict csp.
CONSOLE ERROR: Refused to connect to 'http://127.0.0.1:8000/resources/redirect.php?url=http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php&cors_allow_origin=*&delay=100' because it violates the following Content Security Policy directive: "connect-src 'self'".

CONSOLE ERROR: Refused to connect to 'http://127.0.0.1:8000/resources/redirect.php?url=http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php&cors_allow_origin=*&delay=100' because it violates the document's Content Security Policy.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
This tests the interaction of the fetch API run in the isolated world with the isolated world CSP.
